requirements.txt
blinker==1.9.0
click==8.3.0
Flask==3.1.2
flask-cors==6.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
flask-socketio==5.3.6
python-socketio==5.11.0
eventlet==0.36.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
python-dotenv==1.1.1
SQLAlchemy==2.0.43
stripe==11.2.0
typing_extensions==4.15.0
Werkzeug==3.1.3
